+++ tests/e2e/push_test.go
@@ -3,7 +3,6 @@ package e2e
 import (
 	"bytes"
 	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -14,6 +13,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"helm.sh/helm/v3/pkg/repo"
+
+	"github.com/hypnoglow/helm-s3/internal/helmutil"
 )
 
 const (
@@ -36,15 +37,13 @@ func TestPush(t *testing.T) {
 	setupRepo(t, repoName, repoDir)
 	defer teardownRepo(t, repoName)
 
-	tmpdir, err := os.MkdirTemp("", t.Name())
-	require.NoError(t, err)
-	defer os.RemoveAll(tmpdir)
+	tmpdir := setupTempDir(t)
 
 	// Fetch the repo index before push to get generated time.
 
 	indexFile := filepath.Join(tmpdir, "index.yaml")
 
-	err = mc.FGetObject(repoName, repoDir+"/index.yaml", indexFile, minio.GetObjectOptions{})
+	err := mc.FGetObject(repoName, repoDir+"/index.yaml", indexFile, minio.GetObjectOptions{})
 	require.NoError(t, err)
 
 	idx, err := repo.LoadIndexFile(indexFile)
@@ -297,6 +296,10 @@ func TestPushRelative(t *testing.T) {
 	setupRepo(t, repoName, repoDir)
 	defer teardownRepo(t, repoName)
 
+	tmpdir := setupTempDir(t)
+
+	// Push chart.
+
 	cmd, stdout, stderr := command(fmt.Sprintf("helm s3 push --relative %s %s", chartFilepath, repoName))
 	err := cmd.Run()
 	assert.NoError(t, err)
@@ -304,10 +307,6 @@ func TestPushRelative(t *testing.T) {
 	assert.Contains(t, stdout.String(), "Successfully uploaded the chart to the repository.")
 
 	// Fetch the repo index and check that chart uri is relative.
-
-	tmpdir, err := os.MkdirTemp("", t.Name())
-	require.NoError(t, err)
-	defer os.RemoveAll(tmpdir)
 
 	indexFile := filepath.Join(tmpdir, "index.yaml")
 
@@ -331,7 +330,72 @@ func TestPushRelative(t *testing.T) {
 	err = cmd.Run()
 	assert.NoError(t, err)
 	assertEmptyOutput(t, stdout, stderr)
+
 	assert.FileExists(t, filepath.Join(tmpdir, chartFilename))
+}
+
+func TestPushProvenance(t *testing.T) {
+	t.Log("Test push action for chart with .prov file")
+
+	const (
+		repoName        = "test-push-provenance"
+		repoDir         = "charts"
+		chartName       = "foo"
+		chartVersion    = "1.3.1"
+		chartFilename   = "foo-1.3.1.tgz"
+		provFilename    = chartFilename + ".prov"
+		chartFilepath   = "testdata/" + chartFilename
+		chartObjectName = repoDir + "/" + chartFilename
+		provObjectName  = chartObjectName + ".prov"
+	)
+
+	setupRepo(t, repoName, repoDir)
+	defer teardownRepo(t, repoName)
+
+	tmpdir := setupTempDir(t)
+
+	gnupgHome, err := filepath.Abs("testdata/gnupg")
+	require.NoError(t, err)
+	t.Setenv("GNUPGHOME", gnupgHome)
+
+	// Push chart.
+
+	cmd := fmt.Sprintf("helm s3 push %s %s", chartFilepath, repoName)
+	stdout, stderr, err := runCommand(cmd)
+	assert.NoError(t, err)
+	assertEmptyOutput(t, nil, stderr)
+	assert.Contains(t, stdout.String(), "Successfully uploaded the chart to the repository.")
+
+	// Check that chart was actually pushed.
+
+	obj, err := mc.StatObject(repoName, chartObjectName, minio.StatObjectOptions{})
+	assert.NoError(t, err)
+	assert.Equal(t, chartObjectName, obj.Key)
+
+	// Check that .prov file was actually pushed.
+
+	obj, err = mc.StatObject(repoName, provObjectName, minio.StatObjectOptions{})
+	assert.NoError(t, err)
+	assert.Equal(t, provObjectName, obj.Key)
+
+	// Check that chart can be fetched and verified.
+
+	cmd = fmt.Sprintf("helm fetch %s/%s --version %s --destination %s --verify", repoName, chartName, chartVersion, tmpdir)
+	stdout, stderr, err = runCommand(cmd)
+	assert.NoError(t, err)
+	assertEmptyOutput(t, nil, stderr)
+
+	if helmutil.IsHelm3() {
+		assert.Equal(t, `Signed by: Test Key (helm-s3) <test@example.org>
+Using Key With Fingerprint: 362A31CD7E4CF7E098605C67491EB34640FC8895
+Chart Hash Verified: sha256:be99ea067f7cccf2516cb1e2e43a569f9032d8cde2f28efef200c66662e0b494
+`, stdout.String())
+	} else {
+		assert.Contains(t, stdout.String(), "sha256:be99ea067f7cccf2516cb1e2e43a569f9032d8cde2f28efef200c66662e0b494")
+	}
+
+	assert.FileExists(t, filepath.Join(tmpdir, chartFilename))
+	assert.FileExists(t, filepath.Join(tmpdir, provFilename))
 }
 
 func assertEmptyOutput(t *testing.T, stdout, stderr *bytes.Buffer) {